+++ geo_activity_playground/webui/blueprints/activity_blueprint.py
@@ -214,6 +214,8 @@ def make_activity_blueprint(
             context["heartrate_time_plot"] = heart_rate_time_plot(display_time_series)
         if "cadence" in time_series.columns:
             context["cadence_time_plot"] = cadence_time_plot(display_time_series)
+        if "power" in time_series.columns:
+            context["power_time_plot"] = power_time_plot(display_time_series)
 
         return render_template("activity/show.html.j2", **context)
 
@@ -630,6 +632,20 @@ def cadence_time_plot(time_series: pd.DataFrame) -> str:
         .encode(
             alt.X("time", title=_("Time")),
             alt.Y("cadence", title=_("Cadence")),
+            alt.Color("segment_id:N", title=_("Segment")),
+        )
+        .interactive(bind_y=False)
+        .to_json(format="vega")
+    )
+
+
+def power_time_plot(time_series: pd.DataFrame) -> str:
+    return (
+        alt.Chart(time_series, title=_("Power"))
+        .mark_line()
+        .encode(
+            alt.X("time", title=_("Time")),
+            alt.Y("power", scale=alt.Scale(zero=False), title=_("Power / W")),
             alt.Color("segment_id:N", title=_("Segment")),
         )
         .interactive(bind_y=False)